all-requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
amqp==2.6.1
anyjson==0.3.3
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.1.1
asgiref==3.4.1
astroid==2.6.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
binaryornot==0.4.4
black==21.6b0
bs4==0.0.1
cached-property==1.5.2
cattrs==1.8.0
celery==4.3.0
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==3.2.9
django-allauth==0.44.0
django-appconf==1.0.4
django-celery-beat==1.5.0
django-compressor==2.4.1
django-cors-headers==3.7.0
django-countries==7.2.1
django-coverage-plugin==2.0.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-import-export==2.6.1
django-model-utils==4.1.1
django-pgcrypto==2.0.0
django-redis==5.0.0
django-rest-auth==0.9.5
django-searchable-select==1.5.0
django-stubs==1.8.0
django-stubs-ext==0.2.0
django-timezone-field==4.2.1
django-windows-tools==0.2.1
djangorestframework==3.12.4
djangorestframework-filters==1.0.0.dev0
djangorestframework-simplejwt==4.7.2
docutils==0.17.1
drf-yasg==1.20.0
ed25519==1.5
et-xmlfile==1.1.0
factory-boy==3.2.0
Faker==8.10.0
filelock==3.0.12
flake8==3.9.2
flake8-isort==4.0.0
future==0.18.2
googletrans==4.0.0rc1
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httptools==0.2.0
httpx==0.13.3
hyperframe==5.2.0
identify==2.2.11
idna==2.10
imagesize==1.2.0
IMDbPY==2021.4.18
importlib-metadata==4.6.1
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.2
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jinja2-time==0.2.0
jsonschema==3.2.0
kombu==4.6.11
lazy-object-proxy==1.6.0
libretranslatepy==2.1.1
livereload==2.6.3
lxml==4.6.3
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
multidict==5.1.0
mypy==0.902
mypy-extensions==0.4.3
mysqlclient==2.0.3
nodeenv==1.6.0
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.0
parso==0.8.2
pathspec==0.8.1
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.13.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.9.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
pytest-django==4.4.0
pytest-sugar==0.9.4
python-crontab==2.6.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-filmaffinity==0.0.19
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
rcssmin==1.0.6
redis==3.5.3
regex==2021.7.6
requests==2.25.1
requests-cache==0.8.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rjsmin==1.1.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.0.2
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.25
sqlparse==0.4.1
swagger-spec-validator==2.7.3
tablib==3.1.0
termcolor==1.1.0
testfixtures==6.17.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
translate==3.6.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
url-normalize==1.4.3
urllib3==1.26.6
user-agent==0.1.9
uvicorn==0.14.0
vine==1.3.0
virtualenv==20.4.7
watchgod==0.7
wcwidth==0.2.5
websockets==9.1
Werkzeug==1.0.1
wfastcgi==3.0.0
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.3
zipp==3.5.0
wheel
django-summernote
daphne==3.0.2

